+++ app/controllers/register.py
@@ -9,7 +9,7 @@ from app.helpers.form import create_form, save_form_to_session
 
 from app.models import User
 
-from app.controllers.forms import RegisterForm
+from app.forms import RegisterForm
 
 
 class RegisterView(FlaskView):
