+++ app/forms/diets.py
@@ -3,7 +3,7 @@ from wtforms import validators
 
 from flask_wtf import FlaskForm
 
-from app.controllers.forms.custom import ComaFloatField
+from app.forms.custom import ComaFloatField
 
 
 class DietForm(FlaskForm):
